+++ applovin_max/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Versions
 
+## x.x.x
+    * Add support for ad revenue callbacks
 ## 2.1.0
     * Depends on Android SDK v11.5.2 and iOS SDK v11.5.1.
     * Add API to enable or disable the Creative Debugger `AppLovinMAX.setCreativeDebuggerEnabled(...)`.